feat: public wall feature

+++ src/app/components/wall/wall-notes.tsx
@@ -7,24 +7,35 @@ export type PostType = {
   createdAt: string;
   _id: string;
   isRtl: boolean;
+  author: [
+    { name: string | null | undefined; image: string | null | undefined }
+  ];
 };
 
 interface IWallNotes {
   data: PostType[];
-  removeNoteFromWall: (id: string) => void;
+  removeNoteFromWall?: (id: string) => void;
   setWallNotes: (notes: PostType[]) => void;
-  user: { name: string | null | undefined; image: string | null | undefined };
+  hideOptions?: boolean;
+  hidePrivacy?: boolean;
+  emptyMessage?: string;
+  hideAuthor?: boolean;
 }
 
 export default function WallNotes({
   removeNoteFromWall,
   setWallNotes,
   data,
-  user,
+  hidePrivacy,
+  hideAuthor,
+  hideOptions,
+  emptyMessage,
 }: IWallNotes) {
-  if (data.length === 0) return <EmptyNotes />;
+  if (data.length === 0) return <EmptyNotes emptyMessage={emptyMessage} />;
 
   const deletePost = (id: string) => {
+    if (!removeNoteFromWall)
+      return console.error("removeNoteFromWall is not defined");
     const oldContent = [...data];
     removeNoteFromWall(id);
     fetch(`/api/wall/${id}`, { method: "DELETE" })
@@ -46,26 +57,30 @@ export default function WallNotes({
           <article className="bg-zinc-950/20 p-4">
             <div className="flex justify-between mb-3">
               <div className="flex gap-2 items-center">
-                {user.image ? (
+                {!hideAuthor && note.author[0].image ? (
                   <Image
                     width={32}
                     height={32}
-                    src={user.image}
-                    alt={user.name || "N/A"}
+                    src={note.author[0].image}
+                    alt={note.author[0].name || "N/A"}
                     className="w-8 h-8 rounded-full"
                   />
                 ) : (
                   ""
                 )}
-                <div>{user.name || "N/A"}</div>
+                {hideAuthor ? "" : <div>{note.author[0].name || "N/A"}</div>}
               </div>
               <div>
                 <p className="text-zinc-300 text-sm">
                   {getDateString(note.createdAt)}
                 </p>
-                <p className="text-[0.7rem] text-zinc-500 text-right">
-                  {note.isPublic ? "Public" : "Private"}
-                </p>
+                {hidePrivacy ? (
+                  ""
+                ) : (
+                  <p className="text-[0.7rem] text-zinc-500 text-right">
+                    {note.isPublic ? "Public" : "Private"}
+                  </p>
+                )}
               </div>
             </div>
             <p
@@ -77,18 +92,22 @@ export default function WallNotes({
               {note.content}
             </p>
           </article>
-          <aside className="bg-zinc-950/10 px-4 text-sm py-1 text-zinc-400 rounded-br-xl rounded-bl-xl">
-            <menu className="flex gap-3 justify-end">
-              <li>
-                <button
-                  className="hover:underline"
-                  onClick={() => deletePost(note._id)}
-                >
-                  Delete
-                </button>
-              </li>
-            </menu>
-          </aside>
+          {!hideOptions ? (
+            <aside className="bg-zinc-950/10 px-4 text-sm py-1 text-zinc-400 rounded-br-xl rounded-bl-xl">
+              <menu className="flex gap-3 justify-end">
+                <li>
+                  <button
+                    className="hover:underline"
+                    onClick={() => deletePost(note._id)}
+                  >
+                    Delete
+                  </button>
+                </li>
+              </menu>
+            </aside>
+          ) : (
+            ""
+          )}
         </div>
       ))}
     </>
